refactor to make GCP work
