+++ gcp.go
@@ -1,19 +1,19 @@
-package gcp
+package gcp_trigger
 
 import (
-	"fmt"
-	"log"
-	"net/http"
+	"context"
 	"convictional.com/switchboard/env"
+	"convictional.com/switchboard/triggers/gcp"
 	"convictional.com/switchboard/triggers/shared"
-
+	"fmt"
 	"go.uber.org/zap"
 	"go.uber.org/zap/zapcore"
+	"log"
 )
 
-// httpTriggerEvent is an HTTP Cloud Function with a request parameter.
-func httpTriggerEvent(w http.ResponseWriter, r HTTPWebRequest) {
-
+// TriggerPubSub consumes a Pub/Sub message.
+// https://cloud.google.com/pubsub/docs/reference/rest/v1/PubsubMessage
+func TriggerPubSub(ctx context.Context, gcpPubSub gcp.GCPPubSubRecord) error {
 	config := zap.NewProductionConfig()
 	config.OutputPaths = []string{"stdout"}
 	if env.Debug() {
@@ -24,20 +24,20 @@ func httpTriggerEvent(w http.ResponseWriter, r HTTPWebRequest) {
 	logger, err := config.Build()
 	if err != nil {
 		log.Fatalf("Failed to setup logger :: %+v", err)
-		return
+		return err
 	}
 
-	logger.Debug(fmt.Sprintf("GCP Events :: %+v", r))
+	logger.Debug(fmt.Sprintf("GCP Events :: %+v", gcpPubSub))
 
 	// Convert event to be platform-agnostic
-	event, err := r.ConvertHTTPToTriggerEvent()
+	event, err := gcpPubSub.ConvertPSToTriggerEvent()
 	if err != nil {
 		logger.Error("Failed to convert trigger event", zap.Error(err))
-		return
+		return err
 	}
 
-	service := shared.NewService(logger) // TODO - Move shared drive
+	service := shared.NewService(logger)
 	service.Run(event)
 
-	return
+	return err
 }
